requirements.txt
asgiref==3.7.2
attrs==23.1.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==41.0.2
cssselect2==0.7.0
defusedxml==0.7.1
Django==4.2.6
django-cors-headers==4.2.0
django-extensions==3.2.1
django-jazzmin==2.6.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-spectacular==0.26.5
easy-thumbnails==2.8.1
idna==3.4
inflection==0.5.1
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
lxml==4.9.3
MarkupSafe==2.1.3
oauthlib==3.2.2
Pillow==9.1.0
pycparser==2.21
PyJWT==2.8.0
pyOpenSSL==23.2.0
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.30.2
reportlab==4.0.4
requests==2.28.1
requests-oauthlib==1.3.1
rpds-py==0.10.6
social-auth-app-django==5.0.0
social-auth-core==4.4.2
sqlparse==0.4.4
svglib==1.5.1
tinycss2==1.2.1
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16
webencodings==0.5.1
Werkzeug==2.2.2
